fixed unit test, requires gymnasium>=0.27
gymnasium 0.27 Discrete space returns np.int64 instead of int

+++ nes_py/tests/test_nes_env.py
@@ -83,7 +83,7 @@ class ShouldStepEnv(TestCase):
                 self.assertIsInstance(state, np.ndarray)
             # sample a random action and check it
             action = env.action_space.sample()
-            self.assertIsInstance(action, int)
+            self.assertIsInstance(action, np.int64)
             # take a step and check the outputs
             output = env.step(action)
             self.assertIsInstance(output, tuple)
